ref(/migrations): modify timestamps & modify logic for down

+++ database/migrations/2024_03_18_063500_add_user_id_to_items_table.php
@@ -22,8 +22,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('items', function (Blueprint $table) {
-            $table->dropForeign(['user_id']);
-            $table->dropColumn('user_id');
+            if (Schema::hasColumn('items', 'user_id')) {
+                $table->dropForeign(['user_id']);
+                $table->dropColumn('user_id');
+            }
         });
     }
 };
